# apps/social-network-api/src/app/posts/dto/file-upload.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class FileUploadDto {
    @ApiProperty({ type: 'string', format: 'binary' })
        // tslint:disable-next-line:no-any
    file: any;

    @ApiProperty()
    text: string;
}