# swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterRequestDto"}}}},"responses":{"201":{"description":"User register","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterResponseDto"}}}}},"tags":["auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginRequestDto"}}}},"responses":{"201":{"description":"User login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginResponseDto"}}}}},"tags":["auth"]}},"/auth/validate":{"get":{"operationId":"AuthController_isValidFields","parameters":[],"responses":{"201":{"description":"is valid fields","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IsValidFieldsResponseDto"}}}}},"tags":["auth"]}},"/profile/edit":{"post":{"operationId":"ProfileController_edit","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequestDto"}}}},"responses":{"201":{"description":"Update user profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditProfileResponseDto"}}}}},"tags":["profile"],"security":[{"bearer":[]}]}},"/profile/{userId}":{"get":{"operationId":"ProfileController_get","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get user profile by user id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindProfileResponseDto"}}}}},"tags":["profile"]}},"/profile/editAvatar":{"post":{"operationId":"ProfileController_editAvatar","parameters":[],"responses":{"201":{"description":"Update avatar","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditAvatarResponse"}}}}},"tags":["profile"],"security":[{"bearer":[]}]}},"/profile/editMainImage":{"post":{"operationId":"ProfileController_editMainImage","parameters":[],"responses":{"201":{"description":"Update mainImage","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditMainImageResponseDto"}}}}},"tags":["profile"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UsersController_getUsers","parameters":[],"responses":{"201":{"description":"Get users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetUsersResponseDto"}}}}}},"tags":["users"]}},"/users/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"201":{"description":"Get me","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMeResponseDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get user by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserByIdResponseDto"}}}}},"tags":["users"]}},"/users/follow/{id}":{"get":{"operationId":"UsersController_getFollowUsers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get members","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetFollowUsersResponseDto"}}}}}},"tags":["users"]}},"/follow/{userId}":{"put":{"operationId":"FollowController_follow","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Follow"}},"tags":["follow"]},"delete":{"operationId":"FollowController_unfollow","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Unfollow"}},"tags":["follow"]}},"/posts":{"post":{"operationId":"PostsController_createPost","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileUploadDto"}}}},"responses":{"201":{"description":"Create Post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostResponseDto"}}}}},"tags":["posts"],"security":[{"bearer":[]}]}},"/posts/{parentId}":{"post":{"operationId":"PostsController_createComment","parameters":[{"name":"parentId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileUploadDto"}}}},"responses":{"201":{"description":"Create comment","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommentResponseDto"}}}}},"tags":["posts"],"security":[{"bearer":[]}]}},"/posts/{id}":{"put":{"operationId":"PostsController_updatePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileUploadDto"}}}},"responses":{"201":{"description":"Update post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostResponseDto"}}}}},"tags":["posts"],"security":[{"bearer":[]}]}},"/posts/user/{id}":{"get":{"operationId":"PostsController_getPostsOfUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get user posts by id","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetPostsOfUserResponseDto"}}}}}},"tags":["posts"]}},"/dialogs":{"post":{"operationId":"DialogsController_createDialog","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDialogRequestDto"}}}},"responses":{"201":{"description":"Create dialog","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDialogResponseDto"}}}}},"tags":["dialogs"],"security":[{"bearer":[]}]},"get":{"operationId":"DialogsController_getDialogs","parameters":[],"responses":{"201":{"description":"Get dialogs","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetDialogsResponseDto"}}}}}},"tags":["dialogs"],"security":[{"bearer":[]}]},"put":{"operationId":"DialogsController_updateDialogOwners","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOwnersRequestDto"}}}},"responses":{"201":{"description":"Update owners","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOwnersResponseDto"}}}}},"tags":["dialogs"],"security":[{"bearer":[]}]}},"/dialogs/{id}":{"get":{"operationId":"DialogsController_getDialog","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get dialog","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetDialogResponseDto"}}}}},"tags":["dialogs"],"security":[{"bearer":[]}]}},"/files/{fileName}":{"get":{"operationId":"FilesController_getFile","parameters":[{"name":"fileName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Social network","description":"Social network API ","version":"1.0","contact":{}},"tags":[{"name":"API","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthRegisterRequestDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["firstName","lastName","phone","email","password"]},"AuthRegisterResponseDto":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"AuthLoginRequestDto":{"type":"object","properties":{"loginOrEmail":{"type":"string"},"password":{"type":"string"}},"required":["loginOrEmail","password"]},"AuthLoginResponseDto":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"IsValidFieldsResponseDto":{"type":"object","properties":{}},"UpdateProfileRequestDto":{"type":"object","properties":{"middleName":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"login":{"type":"string"},"birthday":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"relatives":{"type":"string"},"school":{"type":"string"},"status":{"type":"string"},"about":{"type":"string"}},"required":["middleName","firstName","lastName","phone","email","login","birthday","country","city","relatives","school","status","about"]},"EditProfileResponseDto":{"type":"object","properties":{"updated":{"type":"boolean","description":"профиль обновлен","example":true}},"required":["updated"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"login":{"type":"string"}},"required":["id","email","login"]},"FindProfileResponseDto":{"type":"object","properties":{"id":{"type":"string"},"owner":{"default":"ownerId","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"phone":{"type":"string"},"avatar":{"type":"string"},"status":{"type":"string"},"about":{"type":"string"},"birthday":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"relatives":{"type":"string"},"school":{"type":"string"},"mainImage":{"type":"string"}},"required":["id","owner","firstName","lastName","middleName","phone","avatar","status","about","birthday","country","city","relatives","school","mainImage"]},"EditAvatarResponse":{"type":"object","properties":{"fileName":{"type":"string","description":"Название файла","example":"4b890a22-705b-4872-acb6-a8b829ee0e41_avatar_1651603183"}},"required":["fileName"]},"EditMainImageResponseDto":{"type":"object","properties":{"fileName":{"type":"string","description":"Название файла","example":"4b890a22-705b-4872-acb6-a8b829ee0e41_mainImage_1651603183"}},"required":["fileName"]},"GetUsersResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"login":{"type":"string"}},"required":["id","email","login"]},"ProfileEntity":{"type":"object","properties":{"id":{"type":"string"},"owner":{"default":"ownerId","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"firstName":{"type":"string"},"lastName":{"type":"string"},"middleName":{"type":"string"},"phone":{"type":"string"},"avatar":{"type":"string"},"status":{"type":"string"},"about":{"type":"string"},"birthday":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"relatives":{"type":"string"},"school":{"type":"string"},"mainImage":{"type":"string"}},"required":["id","owner","firstName","lastName","middleName","phone","avatar","status","about","birthday","country","city","relatives","school","mainImage"]},"GetMeResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"id пользователя","example":"1b5b029-8bb8-4076-83d8-4f29a58e4dca"},"email":{"type":"string","description":"email пользователя","example":"lev@mail.ru"},"login":{"type":"string","description":"логин пользователя","example":"levnew"},"profile":{"description":"профиль пользователя","allOf":[{"$ref":"#/components/schemas/ProfileEntity"}]}},"required":["id","email","login","profile"]},"GetUserByIdResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"login":{"type":"string"}},"required":["id","email","login"]},"GetFollowUsersResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"id пользователя","example":"1b5b029-8bb8-4076-83d8-4f29a58e4dca"},"email":{"type":"string","description":"email пользователя","example":"lev@mail.ru"},"login":{"type":"string","description":"логин пользователя","example":"levnew"}},"required":["id","email","login"]},"FileUploadDto":{"type":"object","properties":{"file":{"type":"string","format":"binary"},"text":{"type":"string"}},"required":["file","text"]},"CreatePostResponseDto":{"type":"object","properties":{"id":{"type":"string"},"owner":{"$ref":"#/components/schemas/UserEntity"},"text":{"type":"string"},"image":{"type":"string"},"likes":{"type":"number"},"views":{"type":"number"}},"required":["id","owner","text","image","likes","views"]},"CreateCommentResponseDto":{"type":"object","properties":{"id":{"type":"string"},"owner":{"$ref":"#/components/schemas/UserEntity"},"text":{"type":"string"},"image":{"type":"string"},"likes":{"type":"number"},"views":{"type":"number"}},"required":["id","owner","text","image","likes","views"]},"UpdatePostResponseDto":{"type":"object","properties":{"updated":{"type":"boolean","description":"пост обновлен","example":true}},"required":["updated"]},"PostsWithCommentsModel":{"type":"object","properties":{"id":{"type":"string","description":"id поста","example":"d5532bd6-24da-49c3-8e0c-f996f28b7f49"},"text":{"type":"string","description":"текст поста","example":"Всем привет!"},"image":{"type":"string","description":"ссылка на файл","example":"../file/name"},"likes":{"type":"number","description":"количество лайков","example":4},"views":{"type":"number","description":"количество просмотров","example":20},"ownerId":{"type":"string","description":"id владельца поста","example":"e8c4986b-1b5f-47c9-a533-d22329df0da7"},"comments":{"description":"комментарии под постом","type":"array","items":{"$ref":"#/components/schemas/PostsWithCommentsModel"}},"parentPostsId":{"type":"object","description":"под каким постом оставлен данный пост","example":"859d0610-86bd-4733-b954-ab5a491f8e56"}},"required":["id","text","image","likes","views","ownerId","comments","parentPostsId"]},"GetPostsOfUserResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"id поста","example":"d5532bd6-24da-49c3-8e0c-f996f28b7f49"},"text":{"type":"string","description":"текст поста","example":"Всем привет!"},"image":{"type":"string","description":"ссылка на файл","example":"../file/name"},"likes":{"type":"number","description":"количество лайков","example":4},"views":{"type":"number","description":"количество просмотров","example":20},"ownerId":{"type":"string","description":"id владельца поста","example":"e8c4986b-1b5f-47c9-a533-d22329df0da7"},"comments":{"description":"комментарии под постом","type":"array","items":{"$ref":"#/components/schemas/PostsWithCommentsModel"}},"parentPostsId":{"type":"object","description":"под каким постом оставлен данный пост","example":"859d0610-86bd-4733-b954-ab5a491f8e56"}},"required":["id","text","image","likes","views","ownerId","comments","parentPostsId"]},"CreateDialogRequestDto":{"type":"object","properties":{"text":{"type":"string"},"otherOwners":{"type":"array","items":{"type":"string"}}},"required":["text","otherOwners"]},"MessagesEntity":{"type":"object","properties":{"id":{"type":"string"},"dialog":{"default":["dialogId"],"type":"array","items":{"$ref":"#/components/schemas/DialogsEntity"}},"owner":{"default":"owner_id","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"text":{"type":"string"},"image":{"type":"string"},"file":{"type":"string"}},"required":["id","dialog","owner","text","image","file"]},"DialogsEntity":{"type":"object","properties":{"id":{"type":"string"},"owners":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessagesEntity"}}},"required":["id","owners","messages"]},"CreateDialogResponseDto":{"type":"object","properties":{"id":{"type":"string"},"dialog":{"default":["dialogId"],"type":"array","items":{"$ref":"#/components/schemas/DialogsEntity"}},"owner":{"default":"owner_id","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"text":{"type":"string"},"image":{"type":"string"},"file":{"type":"string"}},"required":["id","dialog","owner","text","image","file"]},"GetDialogsResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"id диалога","example":"b92b7fe8-130b-423b-92b8-ba5c2236dfd8"},"userId":{"type":"string","description":"id юзера","example":"e8c4986b-1b5f-47c9-a533-d22329df0da7"},"dialogsId":{"type":"string","description":"id диалога","example":"b92b7fe8-130b-423b-92b8-ba5c2236dfd8"},"lastMessage":{"description":"последние сообщение","type":"array","items":{"type":"array"}}},"required":["id","userId","dialogsId","lastMessage"]},"GetDialogResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"id диалога","example":"b92b7fe8-130b-423b-92b8-ba5c2236dfd8"},"userId":{"type":"string","description":"id юзера","example":"e8c4986b-1b5f-47c9-a533-d22329df0da7"},"dialogsId":{"type":"string","description":"id диалога","example":"b92b7fe8-130b-423b-92b8-ba5c2236dfd8"},"messages":{"description":"сообщения","type":"array","items":{"type":"array"}}},"required":["id","userId","dialogsId","messages"]},"UpdateOwnersRequestDto":{"type":"object","properties":{"dialogId":{"type":"string"},"owners":{"type":"array","items":{"type":"string"}}},"required":["dialogId","owners"]},"UpdateOwnersResponseDto":{"type":"object","properties":{"id":{"type":"string"},"owners":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessagesEntity"}}},"required":["id","owners","messages"]}}}}